Create README.md
Insertion sort is used when the number of elements is small. It can also be useful when the input array is almost sorted; only a few elements are misplaced, incomplete big array. 
What is Binary Insertion Sort?
We can use binary search to reduce the number of comparisons in normal insertion sort. Binary Insertion Sort uses binary search to find the proper location to insert the selected item at each iteration. In normal insertion, sorting takes O(i) (at ith iteration) in the worst case. We can reduce it to O(log) by using binary search. As a whole, the algorithm still has a running worst-case running time of O(n2) because of the series of swaps required for each insertion. Refer to this for implementation.